requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.3
matplotlib>=3.7
opencv-python>=4.8
torch>=2.2
torchvision>=0.17
